refactor tests into multiple files for control in travis, easier editing
